WIP_corrigindo merge

+++ app/Http/Controllers/ChamadoController.php
@@ -8,12 +8,13 @@ use App\Models\Comentario;
 use App\Models\Fila;
 use App\Models\Setor;
 use App\Models\User;
+use App\Models\Patrimonio;
 use App\Utils\JSONForms;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\URL;
-use Illuminate\Support\Facades\Log;
 
 class ChamadoController extends Controller
 {
@@ -21,7 +22,6 @@ class ChamadoController extends Controller
     {
         $this->middleware('auth');
     }
-
     /**
      * Display a listing of the resource.
      *
@@ -30,14 +30,12 @@ class ChamadoController extends Controller
     public function index(Request $request)
     {
         $this->authorize('chamados.viewAny');
-
         if (session('ano') == null) {
             session(['ano' => date('Y')]);
         }
         $chamados = Chamado::listarChamados(session('ano'));
         return view('chamados/index', compact('chamados'));
     }
-
     /**
      * Show the form for creating a new resource.
      *
@@ -53,7 +51,6 @@ class ChamadoController extends Controller
         $form = JSONForms::generateForm($fila);
         return view('chamados/create', compact('fila', 'chamado', 'complexidades', 'status_list', 'form'));
     }
-
     /**
      * Mostra lista de filas e respectivos setores
      * para selecionar e criar novo chamado
@@ -65,10 +62,8 @@ class ChamadoController extends Controller
         $setores = Setor::with(['filas' => function ($query) {
             $query->where('estado', 'Em produção');
         }])->orderBy('sigla')->get();
-
         return view('chamados.listafilas', compact('setores'));
     }
-
     /**
      * Store a newly created resource in storage.
      *
@@ -80,7 +75,6 @@ class ChamadoController extends Controller
         # limitar _as filas que o usuario pode criar, somente filas "em produção"
         $this->authorize('chamados.create');
         $request->validate(JSONForms::buildRules($request, $fila));
-
         # transaction para não ter problema de inconsistência do DB
         $chamado = \DB::transaction(function () use ($request, $fila) {
             $chamado = new Chamado;
@@ -89,11 +83,9 @@ class ChamadoController extends Controller
             $chamado = $this->grava($chamado, $request);
             return $chamado;
         });
-
         $request->session()->flash('alert-info', 'Chamado enviado com sucesso');
         return redirect()->route('chamados.show', $chamado->id);
     }
-
     /**
      * Display the specified resource.
      *
@@ -106,7 +98,6 @@ class ChamadoController extends Controller
         if (!Gate::allows('chamados.view', $chamado)) {
             return view('sem-acesso');
         }
-
         $template = json_decode($chamado->fila->template);
         $extras = json_decode($chamado->extras);
         if (empty($template)) {
@@ -114,24 +105,18 @@ class ChamadoController extends Controller
         }
         $atendentes = $chamado->users()->wherePivot('papel', 'Atendente')->get();
         $autor = $chamado->users()->wherePivot('papel', 'Autor')->first();
-
         # estamos carregando os vinculados diretos.
         # Seria interessante vincular recursivos? Acho que não mas ...
         $vinculados = $chamado->vinculados;
         $complexidades = Chamado::complexidades(true);
         $status_list = Chamado::status(true);
-
         # para o form de adicionar pessoas
         $modal_pessoa['url'] = 'chamados';
         $modal_pessoa['title'] = 'Adicionar observador';
-
         $max_upload_size = env('APP_UPLOAD_MAX_FILESIZE') != null ? ((int) env('APP_UPLOAD_MAX_FILESIZE')) : 16;
-
         $form = JSONForms::generateForm($chamado->fila, $chamado);
-
         return view('chamados/show', compact('atendentes', 'autor', 'chamado', 'extras', 'template', 'vinculados', 'complexidades', 'status_list', 'modal_pessoa', 'max_upload_size', 'form'));
     }
-
     /**
      * Retornando os chamados para criar vinculo entre eles
      * @param Request $request - assunto ou número do chamado
@@ -147,7 +132,6 @@ class ChamadoController extends Controller
                 //busca pelo assunto, independete do ano
                 $chamados = Chamado::listarChamados(null, null, $request->term);
             }
-
             # vamos formatar para datatables
             $results = [];
             foreach ($chamados as $chamado) {
@@ -160,7 +144,6 @@ class ChamadoController extends Controller
         }
         return null;
     }
-
     /**
      * Vincula chamados
      * @param $request->slct_chamados - nrp do chamado a ser vinculado
@@ -169,16 +152,12 @@ class ChamadoController extends Controller
     public function storeChamadoVinculado(Request $request, Chamado $chamado)
     {
         $this->authorize('chamados.create');
-
         if ($request->slct_chamados != $chamado->id) {
-
             $request->validate([
                 'acesso' => ['in:Leitura'],
             ]);
-
             $chamado->vinculadosIda()->detach($request->slct_chamados);
             $chamado->vinculadosIda()->attach($request->slct_chamados, ['acesso' => $request->acesso]);
-
             $vinculado = Chamado::find($request->slct_chamados);
             //comentário no chamado principal
             Comentario::create([
@@ -194,24 +173,20 @@ class ChamadoController extends Controller
                 'comentario' => 'Esse chamado foi vinculado ao chamado no. ' . $chamado->nro . '/' . $chamado->created_at->year,
                 'tipo' => 'system',
             ]);
-
             $request->session()->flash('alert-info', 'Chamado vinculado com sucesso');
         } else {
             $request->session()->flash('alert-warning', 'Não é possível vincular o chamado à ele mesmo');
         }
         return Redirect::to(URL::previous() . "#card_vinculados");
     }
-
     /**
      * Desvincula chamados
      */
     public function deleteChamadoVinculado(Request $request, Chamado $chamado, $id)
     {
         $this->authorize('chamados.create');
-
         $chamado->vinculadosIda()->detach($id);
         $chamado->vinculadosVolta()->detach($id);
-
         $vinculado = Chamado::find($id);
         #dd($vinculado);
         //comentário no chamado principal
@@ -221,7 +196,6 @@ class ChamadoController extends Controller
             'comentario' => 'O chamado no. ' . $vinculado->nro . '/' . $vinculado->created_at->year . ' foi desvinculado desse chamado',
             'tipo' => 'system',
         ]);
-
         // comentário no chamado vinculado
         Comentario::create([
             'user_id' => \Auth::user()->id,
@@ -229,11 +203,9 @@ class ChamadoController extends Controller
             'comentario' => 'Esse chamado foi desvinculado do chamado no. ' . $chamado->nro . '/' . $chamado->created_at->year,
             'tipo' => 'system',
         ]);
-
         $request->session()->flash('alert-info', 'Chamado desvinculado com sucesso');
         return Redirect::to(URL::previous() . "#card_vinculados");
     }
-
     /**
      * Update the specified resource in storage.
      *
@@ -244,38 +216,30 @@ class ChamadoController extends Controller
     public function update(Request $request, Chamado $chamado)
     {
         $this->authorize('chamados.update', $chamado);
-
         # inicialmente atende a atualização do campo anotacoes via ajax
         if ($request->ajax()) {
             $chamado->fill($request->all());
             $chamado->save();
             return response()->json(['message' => 'success', 'data' => $chamado]);
         }
-
         # acho que valida atendente
         if (Gate::allows('admin') and isset($request->atribuido_para)) {
             $request->validate([
                 'fila_id' => ['required|numeric'],
             ]);
         }
-
         # valida customforms
         # está dando erro
         //$request->validate(JSONForms::buildRules($request, $chamado->fila));
-
         $chamado = $this->grava($chamado, $request);
-
         /*
         if(config('app.env') == 'production')
         Mail::send(new ChamadoMail($chamado,$user));
-
         depois de atualizar, tem de registrar nos comentários
          */
-
         $request->session()->flash('alert-info', 'Chamado enviado com sucesso');
         return redirect()->route('chamados.show', $chamado->id);
     }
-
     /**
      * Remove the specified resource from storage.
      *
@@ -286,11 +250,9 @@ class ChamadoController extends Controller
     {
         //
     }
-
     /* Evita duplicarmos código  */
     private function grava(Chamado $chamado, Request $request)
     {
-
         if ($request->status == 'devolver') {
             $chamado->status = 'Triagem';
             $chamado->complexidade = null;
@@ -330,7 +292,6 @@ class ChamadoController extends Controller
                     $chamado->complexidade = $request->complexidade;
                 }
             }
-
             /* Caso tenha alguma atualização, guarda nos registros */
             if (count($atualizacao) > 0) {
                 if (count($atualizacao) == 1) {
@@ -351,7 +312,6 @@ class ChamadoController extends Controller
                     'tipo' => 'system',
                 ]);
             }
-
             /* Administradores */
             if (Gate::allows('admin')) {
                 /* trocar requisitante */
@@ -364,7 +324,6 @@ class ChamadoController extends Controller
                 } else {
                     $user = \Auth::user();
                 }
-
                 /* Atribuir */
                 if (!empty($request->atribuido_para)) {
                     die('desativado');
@@ -376,7 +335,6 @@ class ChamadoController extends Controller
             } else {
                 $user = \Auth::user();
             }
-
             /* Atualiza telefone da pessoa */
             /* para funcionar, precisa mexer no LoginController */
             $user->telefone = $request->telefone;
@@ -388,7 +346,6 @@ class ChamadoController extends Controller
         }
         return $chamado;
     }
-
     /**
      * adiciona atendentes. Pode ser mais de um
      */
@@ -401,28 +358,23 @@ class ChamadoController extends Controller
             return Redirect::to(URL::previous() . "#card_atendente");
         }
         $atendente = User::obterPorCodpes($request->codpes);
-
         # Se atendente já existe não vamos adicionar novamente
         if ($chamado->users()->where(['user_id' => $atendente->id, 'papel' => 'Atendente'])->exists()) {
             $request->session()->flash('alert-info', 'Atendente já existe');
             return Redirect::to(URL::previous() . "#card_atendente");
         }
-
         $chamado->users()->attach($atendente->id, ['papel' => 'Atendente']);
         $chamado->status = 'Atribuído';
         $chamado->save();
-
         Comentario::create([
             'user_id' => \Auth::user()->id,
             'chamado_id' => $chamado->id,
             'comentario' => 'O chamado foi atribuído para o(a) atendente ' . $atendente->name,
             'tipo' => 'system',
         ]);
-
         $request->session()->flash('alert-info', 'Atendente adicionado com sucesso');
         return Redirect::to(URL::previous() . "#card_atendente");
     }
-
     /**
      * Salva o ano na sessão usada no index
      */
@@ -434,7 +386,6 @@ class ChamadoController extends Controller
         }
         return back();
     }
-
     /**
      * Adicionar pessoas relacionadas ao chamado
      * autorizado a qualquer um que tenha acesso ao chamado
@@ -442,6 +393,8 @@ class ChamadoController extends Controller
      */
     public function storePessoa(Request $request, Chamado $chamado)
     {
+        $this->authorize('chamados.view', $chamado);
+
         $user = User::obterOuCriarPorCodpes($request->codpes);
         $chamado->users()->attach($user, ['papel' => 'Observador']);
 
@@ -451,12 +404,9 @@ class ChamadoController extends Controller
             'comentario' => 'O observador ' . $user->name . ' foi adicionado ao chamado.',
             'tipo' => 'system',
         ]);
-
         $request->session()->flash('alert-info', 'Observador adicionado com sucesso.');
-
         return Redirect::to(URL::previous() . "#card_pessoas");
     }
-
     /**
      * Remove pessoas relacionadas ao chamado
      * Autorização: se for remover autor, ou atendente,somente para atendente ou admin
@@ -465,15 +415,17 @@ class ChamadoController extends Controller
      */
     public function destroyPessoa(Request $request, Chamado $chamado, User $user)
     {
+        $this->authorize('chamados.view', $chamado);
+
         $papel = $chamado->users()->where('users.id', $user->id)->first()->pivot->papel;
         $chamado->users()->detach($user);
+        $chamado->users()->wherePivot('papel', $papel)->detach($user);
 
         # verificar se sobrou algum atendente, se não, muda o status
         if (!count($chamado->users()->wherePivot('papel', 'Atendente')->get())) {
             $chamado->status = 'Triagem';
             $chamado->save();
         }
-
         $msg = 'O ' . strtolower($papel) . ' ' . $user->name . ' foi removido desse chamado.';
         Comentario::create([
             'user_id' => \Auth::user()->id,
@@ -482,98 +434,44 @@ class ChamadoController extends Controller
             'tipo' => 'system',
         ]);
         $request->session()->flash('alert-info', $msg);
-
         return Redirect::to(URL::previous() . "#card_pessoas");
     }
-
     /**
-     *
-     * desativados
-     *
-     *
-     *
-     *
+     * Adicionar patrimonios relacionadas ao chamado
+     * autorizado a qualquer um que tenha acesso ao chamado
+     * request->codpes = required, int
      */
-
-    /**
-     * **** Passou a ser feito via modal ****
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Chamado  $chamado
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Chamado $chamado)
+    public function storePatrimonio(Request $request, Chamado $chamado)
     {
-        $this->authorize('chamados.view', $chamado);
-        $fila = $chamado->fila;
-        $atendentes = [];
-        $complexidades = Chamado::complexidades();
-        $autor = $chamado->users()->wherePivot('papel', 'Autor')->first();
-        $form = JSONForms::generateForm($fila, $chamado);
-        return view('chamados/edit', compact('fila', 'chamado', 'atendentes', 'form', 'complexidades'));
-    }
-
-    public function devolver(Chamado $chamado)
-    {
-        return "desativado";
-        $this->authorize('atendente');
-        return view('chamados/devolver', compact('chamado'));
-    }
-
-    public function todos(Request $request)
-    {
-        return "desativado";
-        $this->authorize('admin');
-        $chamados = Chamado::orderBy('created_at', 'desc');
-        // search terms
-        if (isset($request->status)) {
-            $chamados->where('status', '=', $request->status);
-        }
-        if (isset($request->search)) {
-            $chamados->where('chamado', 'LIKE', "%" . $request->search . "%");
-        }
-        $chamados = $chamados->paginate(10);
-        return view('chamados/todos', compact('chamados'));
-    }
-
-    public function buscaid(Request $request)
-    {
-        return 'desativado';
-        $this->authorize('atendente');
-        $chamado = isset($request->id) ? Chamado::find($request->id) : null;
-        $mensagem = null;
-        if (isset($request->id) and is_null($chamado)) {
-            $mensagem = 'Não há chamado com este Id.';
-        }
-        return view('chamados/buscaid', compact('chamado', 'mensagem'));
-    }
-
-    # acho que nao vai mais usar
-    public function triagem()
-    {
-        return 'disable';
-        /* Chamados de quem está logado */
-        $this->authorize('chamados.viewAny');
-        $user = \Auth::user();
-        $chamados = Chamado::where('status', 'Triagem')->orderBy('created_at', 'desc')->get();
-        return view('chamados/index', compact('chamados'));
-    }
-
-    public function atender()
-    {
-        return 'disable';
-        /* Chamados de quem está logado */
-        $this->authorize('chamados.viewAny');
-
-        $user = \Auth::user();
-        $chamados = Chamado::whereHas('users', function ($pivot) {
-            $user = \Auth::user();
-            $pivot->where([
-                ['user_id', $user->id],
-                ['papel', 'Atendente'],
+        if (config('chamados.usar_replicado') == 'true') {
+            $request->validate([
+                'numpat' => 'required|patrimonio',
             ]);
-        })->orderBy('created_at', 'desc')->paginate(10);
+        } else {
+            $request->validate([
+                'numpat' => 'required',
+            ]);
+        }
+        $patrimonio = Patrimonio::where('numpat', $request->numpat)->first();
+        if (!$patrimonio) {
+            $patrimonio = new Patrimonio;
+            $patrimonio->numpat = $request->numpat;
+            $patrimonio->save();
+        }
+        $chamado->patrimonios()->detach($patrimonio);
+        $chamado->patrimonios()->attach($patrimonio);
 
-        return view('chamados/index', compact('chamados'));
+        # continua ...
+        return Redirect::to(URL::previous() . "#card_patrimonios");
+    }
+
+    /**
+     * Remove patrimonios relacionadas ao chamado
+     * Autorização: se for remover autor, ou atendente,somente para atendente ou admin
+     * se for observador, qualquer um que tenha acesso ao chamado
+     * $user = required
+     */
+    public function destroyPatrimonio(Request $request, Chamado $chamado, User $user)
+    {
     }
 }
